Added individual server responsiveness control
Now if $response is called, it is server specific s.t. it is not a master kill switch (previously, if one server did $responding false then all responding values will be false and so the bot isn't available for any server - obviously it was a very weird implementation by me but I just wanted a general idea of what I wanted do). This is achieved by adding database values for each specific server on line 21 using msg.guild.id, and initializing this value to be true. Now, as seen on line 42,